requirements.txt
fitz
spacy
requests
python-dotenv